Move some protocol definitions to their own source files
