+++ Sources/Helpers/PusherEventFactory.swift
@@ -1,9 +1,6 @@
 import Foundation
 
-protocol PusherEventFactory {
-
-    func makeEvent(fromJSON json: PusherEventPayload, withDecryptionKey decryptionKey: String?) throws -> PusherEvent
-}
+// MARK: - Types
 
 // MARK: - Concrete implementation
