+++ Sources/Protocols/EventFactory.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol EventFactory {
+
+    func makeEvent(fromJSON json: PusherEventPayload, withDecryptionKey decryptionKey: String?) throws -> PusherEvent
+}
