+++ Sources/Protocols/EventQueue.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+protocol EventQueue {
+
+    var delegate: EventQueueDelegate? { get set }
+
+    func enqueue(json: PusherEventPayload)
+}
